requirements.txt
flask
pymongo
qrcode
pillow
python-dotenv
pyzbar 